Add files via upload
This project was complicated for me since I could not figure out how to string more than on constant together. After looking on youtube and seeing the source code and comparing what was wrong with mine I found that the original poster was doing the same thing I was. By creating a constant that could be called upon it provided the function and display I was looking for. This was fun